+++ src/apps/DDBCharacterManager.js
@@ -1368,6 +1368,7 @@ ${item.system.description.chat}
       await createInfusedItems(this.ddbCharacter.source.ddb, this.actor);
 
       await this.ddbCharacter.autoLinkResources();
+      await this.ddbCharacter.fixItemSpellResources();
 
       if (this.settings.useChrisPremades) {
         await ExternalAutomations.addChrisEffectsToActorDocuments(this.actor);