+++ src/parser/item/weapon.js
@@ -277,7 +277,7 @@ export default function parseWeapon(data, character, flags) {
   if (flags.classFeatures.includes("OffHand")) weapon.system.activation.type = "bonus";
 
   weapon.system.range = getRange(data, weapon.system.properties);
-  weapon.system.uses = getUses(data);
+  weapon.system.uses = getUses(data, false);
   // force weapons to always not use prompt
   weapon.system.uses.prompt = false;
   weapon.system.ability = "";